+++ changelog.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Removed
 
 ### Fixed
+- fix "Users have to once again Clock In when ViewTouch exits and is restarted" #108
 
 
 ## [v19.05.1] - 2019-05-26